+++ pkg/providers/clientlists/testdata/active_list/imports.sh
@@ -1,0 +1,4 @@
+terraform init
+terraform import akamai_clientlist_list.list_123_ABC 123_ABC
+terraform import akamai_clientlist_activation.activation_123_ABC_PRODUCTION 123_ABC:PRODUCTION
+terraform import akamai_clientlist_activation.activation_123_ABC_STAGING 123_ABC:STAGING