+++ pkg/providers/clientlists/testdata/active_list/variables.tf
@@ -1,0 +1,19 @@
+variable "edgerc_path" {
+  type    = string
+  default = "edgerc_path"
+}
+
+variable "config_section" {
+  type    = string
+  default = "test_section"
+}
+
+variable "contract_id" {
+  type    = string
+  default = "12_CA"
+}
+
+variable "group_id" {
+  type    = number
+  default = 12
+}