+++ pkg/providers/cloudwrapper/testdata/multiline_comment/import.sh
@@ -1,0 +1,2 @@
+terraform init
+terraform import akamai_cloudwrapper_configuration.test_configuration 12345